Simplify macOS icon generation
In hindsight there is no good reason to keep the icon files in a naming convention that does not fit what iconutil expects. Sure, we save a few kilobytes of disk space, but at the expense of adding a potential breaking point in the build process.
This also removes some utilities that are obsolete.